Add files via upload
a script with calculation of WUE using linear and full model.
Mostly: input for tree-rings data and  13C and concentration of atmospheric CO2.
Also, detailed notes on model parametrization from Keeling et al 2017, Seibt et al, and others. Tracked down parametars values as described in these various publication in order to run the full model (as described in Keeling supplementary material).
This part can be cleaned and inserted in the main script that Juergen set up.